# satelite-catalogo-backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.CreateUserRequest:
    properties:
      cc:
        items:
          type: integer
        type: array
      password:
        type: string
      rut:
        type: string
      username:
        type: string
    type: object
...
  models.Role:
    enum:
    - Usuario
    - Administrador
    type: string
    x-enum-varnames:
    - USER
    - ADMIN
...
  models.User:
    properties:
      _id:
        type: string
      cc:
        items:
          type: string
        type: array
      created_at:
        type: string
      email:
        type: string
      role:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      rut:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Catalogo service.
  title: Catalogo API
  version: "1.0"
paths:
...
  /user/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user after validating credentials
      parameters:
      - description: User info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - users
  /user/{email}:
    delete:
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete user
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - description: User info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - users
  /user/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - users
  /user/by-cc:
    post:
      consumes:
      - application/json
      parameters:
      - description: CC IDs
        in: body
        name: body
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get users by centros de costo
      tags:
      - users
  /user/email/{email}:
    get:
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get user by email
      tags:
      - users
swagger: "2.0"